+++ packages/stream_chat/lib/src/core/models/member.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: deprecated_member_use_from_same_package
+
 import 'package:equatable/equatable.dart';
 import 'package:json_annotation/json_annotation.dart';
 import 'package:stream_chat/src/core/models/user.dart';
@@ -15,6 +17,7 @@ class Member extends Equatable {
     this.inviteRejectedAt,
     this.invited = false,
     this.role,
+    this.channelRole,
     this.userId,
     this.isModerator = false,
     DateTime? createdAt,
@@ -46,7 +49,11 @@ class Member extends Equatable {
   final bool invited;
 
   /// The role of the user in the channel
+  @Deprecated('Please use channelRole')
   final String? role;
+
+  /// The role of this member in the channel
+  final String? channelRole;
 
   /// The id of the interested user
   final String? userId;
@@ -76,6 +83,7 @@ class Member extends Equatable {
     DateTime? inviteRejectedAt,
     bool? invited,
     String? role,
+    String? channelRole,
     String? userId,
     bool? isModerator,
     DateTime? createdAt,
@@ -93,6 +101,7 @@ class Member extends Equatable {
         banExpires: banExpires ?? this.banExpires,
         shadowBanned: shadowBanned ?? this.shadowBanned,
         role: role ?? this.role,
+        channelRole: channelRole ?? this.channelRole,
         userId: userId ?? this.userId,
         isModerator: isModerator ?? this.isModerator,
         createdAt: createdAt ?? this.createdAt,
@@ -109,6 +118,7 @@ class Member extends Equatable {
         inviteRejectedAt,
         invited,
         role,
+        channelRole,
         userId,
         isModerator,
         banned,